Move logic from Board component to Game component
The Board component only needs to know what to render. The logic in
the Game component now decides what to tell Board to render and when.
This is optimal because now we maintain state in only one component.

This gives us more control over the state of the game and will also make
testing a lot easier.