Add incorrectLetter counter
So as to not get into an infinite setState loop we check for incorrect
letters only on keyUp events and don’t allow backspaces to affect
the incorrectLetter count. We have to do this on keyUp events rather
than onChange or keyDown because the TypingBox component implementation
requires us to check for an incorrectLetter after keyboard input has
been set, else it won’t count single mistakes.