Calculate WPM more accurately
In this implementation of a typing test you are forced to fix your
typing mistakes so I would argue that it is unreasonable to factor
mistakes into the WPM calculation. Per documentation on calculating
WPM (https://www.speedtypingonline.com/typing-equations) we factor in
an average word length of 5 to calculate gross WPM.

Fixing mistakes not only messes up the players rhythm and likely causes
brief pauses, but function keys like the backspace key take time to
press and do not count as a keyed entry. This means that the time
penalty is already built into the process of correcting a mistake.