End game on final letter
The user no longer needs to press a space to end the game. This is
an intuitive way for the game to end. to justify what's happening here,
we figure out that the last letter has been typed correctly so we need
to advance the game forward by one word causing the game to render in
the end game state, causing the statistics to appear and the Race
component to render at the 'finish line'.

Normally this end game state would be triggered by pressing space on
the final word, we are basically breaking into the end game state before
that happens.