Implement a WebsocketController component
By utilizing a Component above the Game component that uses the
react-websocket library we can effectively control the state of the
game around our websocket connection.

We render the Game component and pass in the ability to make race
updates as props so that any changes to the race can bubble up and be
passed over the websocket connection to the server.

We let the Game component know what its ID is so it can make rendering
decisions based on this variable.