Re-factor Race component and add Player component
The Race component needs the ability to render Player components in the
race so that we can render Players in the game. We also need the
ability to update the Race based on updates from the websocket so that
we can keep the visuals accurate to the race status.

The Player component separates some responsibility from the Race
component so that we can deal with the actual positioning of a Player
in the race.