Add stats page
By hitting an api endpoint to build a stats payload for a player we can
display their global stats. We can then use this information and some
sort of Javascript graph library to build graphs.

I would like to put the stats page into some sort of navbar at the top.

At the moment we just use a flex box to render all game history and wrap
accordingly but it would be nice to put the games into some sort of
scrolling box and display them as rows.

I also need to figure out how to test API endpoints in React. The other
files that utilize endpoints need tests as well.