Implement Graph component
This is a first iteration of what this could look like. I'm confident a
lot of this code can be separated into separate files in to make this
look prettier.

There are a few nested loops happening that might get slow with massive
data sets. Averaging isn't inherently a recursive problem but 1000+
games down the line we might want to consider programming this
dynamically, storing the results of sub-problems in local storage so
that we don't need to compute them as new games are played and the graph
is remade.

Since we are computing the average wpm, accuracy, and mistakes, at each
game played, we are having to perform 3 separate loops inside of the
for loop. This is because the Array reduce() method is a for loop
itself. To be specific, This is an O(n^2) problem that will make the
graph load slower over time. ¯\_(ツ)_/¯