Rate limit position updates and race renders
We can use state to keep track of our old position rather than checking
for it. We can ensure that we only send our new position to the server
when we've progressed 10%.

I spent the day trying to figure out different ways and tricks in React
to achieve all of this and was only semi-successful. I've opted to ask
the server every 2 seconds for a race update. I'm not entirely sure
how I would go about achieving this on the server without having the
client ask for it because some sort of game loop would lock up the
Controller to the multitude of other requests being processed.

So, we can start making requests every 2 seconds when the game has
started, and stop when the game is over. While that is happening we
only send our updates every 10%. It's worth a shot.