Add max character limit to typing box
This will stop players from making a mistake and then continuing to type
limitless characters, racking up their mistakes, before they notice
that they are typing incorrectly.

Having a hard-coded value means that we need to make sure we don't
introduce words that are larger than 11 characters though.

An alternative to this could be introduced when we pull text from the
server, that the server also send over a proper max character count.

Another alternative would be to find the largest string in the text and
add a bunch to its length and set that as the max.

For the sake of demo-ing this, I will just hard code a max length and
ensure that I don't add words that are too long.