Follow lifecycle pattern for API calls
Making this API call is something we want to do as soon as we can. By
placing it in a lifecycle block we make this API call after the first
render any time you navigate to this page if you aren't already
authenticated. This call can only be made after you have signed in
with Slack as seen by the conditional call.

This ensures that this call is only happening once every time you
access the page, the first time it gets turned away, the second time
(after OAuth with Slack) the call goes through.