Add cleanup for timer
This bug has existed for months but only now do I understand how to
fix it. Basically an update was trying to be performed on a component
that was unmounted (this timer) and we needed to clear the timeout
when the timer was finished so that it doesn't try to update.